+++ CHANGELOG.md
@@ -12,6 +12,7 @@ backwards compatibility.
 * Visualize selected scale mode on module's display.
 * Visualize selected chord root on module's display.
 * Visualize selected wavetable bank on module's display.
+* Visualize selected wavetable on module's display.
 
 ## 0.6.0